+++ src/Views/sqlcsstheme.tpl.php
@@ -4,14 +4,8 @@
      * SQL CSS Theme
      * -------------------------------------------------------------
      * 
-     * Configurations for light and dark mode.
-     * Set CSS theme name WITHOUT css extension. E.g 'github'.
-     * 
-     * To get available CSS themes in controller use
-     *     service('highlighter')->getAvailableStyleSheets(); 
-     * or to get available CSS theme with absolute path use
-     *     service('highlighter')->getAvailableStyleSheets(true);
-     * 
+     * Configuration for light and dark mode SQL syntax highlighter.
+     *
      * @var array
      */
     public $sqlCssTheme = [